habit_tracker/requirements.txt
Django>=5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3.1
django-cors-headers>=4.3.0
django-filter>=23.5
psycopg2-binary>=2.9.9
python-decouple>=3.8
celery>=5.3.0
redis>=5.0.0
requests>=2.31.0
drf-spectacular>=0.27.0